+++ packages/studio/core/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.121](https://github.com/andremichelle/openDAW/compare/@opendaw/studio-core@0.0.120...@opendaw/studio-core@0.0.121) (2026-03-07)
+
+**Note:** Version bump only for package @opendaw/studio-core
+
+## [0.0.120](https://github.com/andremichelle/openDAW/compare/@opendaw/studio-core@0.0.119...@opendaw/studio-core@0.0.120) (2026-03-06)
+
+**Note:** Version bump only for package @opendaw/studio-core
+
 ## [0.0.119](https://github.com/andremichelle/openDAW/compare/@opendaw/studio-core@0.0.118...@opendaw/studio-core@0.0.119) (2026-03-04)
 
 **Note:** Version bump only for package @opendaw/studio-core